add Data class and most some things over from App
